# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenClaw 技能平台 API
  description: |
    OpenClaw 技能平台 RESTful API 文档

    提供技能管理、用户管理、搜索、版本控制和审核等功能。
  version: 1.0.0
  contact:
    name: OpenClaw 技术支持
    email: support@openclaw.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api
    description: 本地开发环境
  - url: https://api.openclaw.com/v1
    description: 生产环境
...
paths:
...
  /skills:
    get:
      tags:
        - 技能
      summary: 获取技能列表
      description: 分页获取技能列表，支持分类筛选和排序
      operationId: getSkills
      parameters:
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: category
          in: query
          description: 按分类筛选
          schema:
            type: string
            enum: [productivity, automation, ai, development, integration, utility]
        - name: sort_by
          in: query
          description: 排序方式
          schema:
            type: string
            enum: [popular, newest, downloads, rating]
            default: popular
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillListResponse'

    post:
      tags:
        - 技能
      summary: 创建技能
      description: 发布新技能到平台
      operationId: createSkill
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateSkillRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /skills/{skill_id}:
    get:
      tags:
        - 技能
      summary: 获取技能详情
      description: 获取指定技能的详细信息
      operationId: getSkill
      parameters:
        - $ref: '#/components/parameters/SkillId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - 技能
      summary: 更新技能
      description: 更新技能信息
      operationId: updateSkill
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SkillId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSkillRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags:
        - 技能
      summary: 删除技能
      description: 删除指定技能
      operationId: deleteSkill
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SkillId'
      responses:
        '204':
          description: 删除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/skills:
    get:
      tags:
        - 审核
      summary: 获取待审核技能
      description: 获取技能审核列表（管理员权限）
      operationId: getPendingSkills
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: 审核状态
          schema:
            type: string
            enum: [pending, approved, rejected]
            default: pending
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SkillId:
      name: skill_id
      in: path
      description: 技能 ID
      required: true
      schema:
        type: string
...
  schemas:
...
    Skill:
      type: object
      properties:
        id:
          type: string
          description: 技能 ID
        name:
          type: string
          description: 技能名称
        description:
          type: string
          description: 技能描述
        author:
          type: string
          description: 作者
        version:
          type: string
          description: 版本号
        category:
          type: string
          enum: [productivity, automation, ai, development, integration, utility]
          description: 分类
        tags:
          type: array
          items:
            type: string
          description: 标签
        rating:
          type: number
          format: float
          description: 评分
        downloads:
          type: integer
          description: 下载量
        status:
          type: string
          enum: [published, draft, archived]
          description: 状态
        created_at:
          type: string
          format: date-time
          description: 创建时间
        updated_at:
          type: string
          format: date-time
          description: 更新时间
    SkillDetail:
      allOf:
        - $ref: '#/components/schemas/Skill'
        - type: object
          properties:
            readme:
              type: string
              description: README 内容
            changelog:
              type: string
              description: 变更日志
            versions:
              type: array
              items:
                $ref: '#/components/schemas/SkillVersion'
              description: 版本历史
    SkillVersion:
      type: object
      properties:
        version:
          type: string
          description: 版本号
        created_at:
          type: string
          format: date-time
          description: 创建时间
        changelog:
          type: string
          description: 变更说明
        author:
          type: string
          description: 作者
    CreateSkillRequest:
      type: object
      required:
        - name
        - description
        - category
        - version
        - archive
      properties:
        name:
          type: string
          description: 技能名称
        description:
          type: string
          description: 技能描述
        category:
          type: string
          enum: [productivity, automation, ai, development, integration, utility]
          description: 分类
        tags:
          type: array
          items:
            type: string
          description: 标签
        version:
          type: string
          description: 版本号
        archive:
          type: string
          format: binary
          description: 技能包文件
    UpdateSkillRequest:
      type: object
      properties:
        name:
          type: string
          description: 技能名称
        description:
          type: string
          description: 技能描述
        category:
          type: string
          enum: [productivity, automation, ai, development, integration, utility]
          description: 分类
        tags:
          type: array
          items:
            type: string
          description: 标签
        version:
          type: string
          description: 版本号
...
    SkillListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        total:
          type: integer
          description: 总数量
        page:
          type: integer
          description: 当前页码
        page_size:
          type: integer
          description: 每页数量
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: 错误详情
      required:
        - detail
  responses:
    BadRequest:
      description: 请求错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: 验证错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'